+++ src/main/java/io/github/ramiro/escapesj/servicio/ItemFacturacion.java
@@ -1,9 +1,11 @@
 package io.github.ramiro.escapesj.servicio;
 
+import java.math.BigDecimal;
+
 public record ItemFacturacion(
         String tipo,
         String descripcion,
         String codigoProducto,
         int cantidad,
-        double precioUnitario
+        BigDecimal precioUnitario
 ) {}
